Added conditional code block to Write
This conditional code block is specificaly for overloads.
Sphinx is not including overload method from parent class.

My current work around is to include them in the child class,
Write in this case.

The included overload methods are in a
if mock_g.DOCS_BUILDING: code block.
This means the overload method are only seen by Write
when docs are being built.